requirements.txt
monitorcontrol>=4.2.0
pystray>=0.19.5
Pillow>=10.0.0
keyboard>=0.13.5
schedule>=1.2.0
customtkinter>=5.2.0
astral>=3.2
